+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.3.0] - 2023-12-11
+
+### Added
+
+- New `--latest` flag to `sedge generate` subcommands to use the latest version of the image in the clients.
+- Integrate Holesky network.
+
+### Fixed
+
+- Remove TTD of the params, test cases, command flags and geth patch for custom testnets.
+
+### Changed
+
+- Remove default bootnodes for networks managed by clients.
+- Remove custom configurations for Chiado and instead use the clients configuration.
+- Update client images.
+
 ## [v1.2.3] - 2023-08-16
 
 ### Added
@@ -109,8 +126,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename sedge data directory name to `sedge-data` instead of `docker-compose-scripts`.
 - Update mev-boost relay URLs.
 - Fixed validator restart failure. Validator never restarts, but has a better and safer start-up method:
-    - Validator waits a grace period (2 epochs) before starting.
-    - Validator waits for the consensus sync to finish: verifying that the `/eth/v1/node/health` consensus endpoint
+  - Validator waits a grace period (2 epochs) before starting.
+  - Validator waits for the consensus sync to finish: verifying that the `/eth/v1/node/health` consensus endpoint
       returns with the HTTP code 200 in the newly added `validator-blocker` docker-compose service. This replace the
       older track sync method.
 - Updated installation script for docker and docker compose in Linux flavours.